+++ src/calc/__init__.py
@@ -1,0 +1,12 @@
+"""
+Various calculation challenges
+==============================
+
+"""
+
+from calc.func import *
+
+__all__ = [
+    "get_fibonacci_number",
+    "get_sum_of_strings",
+]
