+++ tests/sequences/test_func.py
@@ -55,15 +55,6 @@ def test_parentheses_validator():
     assert sequences.are_parentheses_balanced("(some <text>)") is True
 
 
-def test_fibonacci_number_getter():
-    assert sequences.get_fibonacci_number(-10) == 0  # special test case
-    assert sequences.get_fibonacci_number(0) == 0
-    assert sequences.get_fibonacci_number(1) == 1
-    assert sequences.get_fibonacci_number(2) == 1
-    assert sequences.get_fibonacci_number(3) == 2
-    assert sequences.get_fibonacci_number(9) == 34
-
-
 # noinspection SpellCheckingInspection
 def test_get_longest_uniq_sequence_length():
     assert sequences.get_longest_uniq_length("abcdefg") == 7
@@ -71,27 +62,4 @@ def test_get_longest_uniq_sequence_length():
     assert sequences.get_longest_uniq_length("hwccjayhiszbmomlqkem") == 11
 
 
-def test_sum_of_strings():
-    assert sequences.get_sum_of_strings("123", "456") == "579"
-    assert sequences.get_sum_of_strings("123", "789") == "912"
-    assert sequences.get_sum_of_strings("99", "1") == "100"
 
-
-def test_sum_of_strings_big_numbers():
-    x, y, test_value = "490053634", "20846201", "510899835"
-    assert sequences.get_sum_of_strings(x, y) == test_value
-
-    x, y, test_value = "616526149", "364644330", "981170479"
-    assert sequences.get_sum_of_strings(x, y) == test_value
-
-
-def test_sum_of_strings_leading_zeros():
-    assert sequences.get_sum_of_strings("009", "000001") == "10"
-    assert sequences.get_sum_of_strings("020", "000000") == "20"
-    assert sequences.get_sum_of_strings("000", "000000") == "0"
-
-
-def test_sum_of_strings_empty():
-    assert sequences.get_sum_of_strings("123", "") == "123"
-    assert sequences.get_sum_of_strings("", "456") == "456"
-    assert sequences.get_sum_of_strings("", "") == "0"
